LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigate-event/defaultPrevented-navigation-preempted-expected.txt


PASS navigateEvent.defaultPrevented isn't affected when the navigation is preempted after dispatch
